feat: added front for notification page
