+++ app/profile/routes.py
@@ -1,10 +1,8 @@
-import json
 import secrets
 import os
 from datetime import datetime, timezone
 
 from app.config import UserConfig
-
 
 from flask import (
     Blueprint,
@@ -716,51 +714,3 @@ def liked_by_me():
         (current,),
     )
     return jsonify([dict(r) for r in rows])
-
-
-@profile_bp.route("/notifications", methods=["GET"])
-@login_required
-def list_notifications():
-    current = g.current_user["id"]
-    unread_only = request.args.get("unread") == "1"
-    if unread_only:
-        rows = query_all(
-            "SELECT * FROM notifications WHERE user_id = ? AND is_read = 0 ORDER BY id DESC LIMIT 100",
-            (current,),
-        )
-    else:
-        rows = query_all(
-            "SELECT * FROM notifications WHERE user_id = ? ORDER BY id DESC LIMIT 100",
-            (current,),
-        )
-
-    payload = []
-    for row in rows:
-        item = dict(row)
-        if item.get("payload"):
-            try:
-                item["payload"] = json.loads(item["payload"])
-            except Exception:
-                pass
-        payload.append(item)
-
-    return jsonify(payload)
-
-
-@profile_bp.route("/notifications/unread-count", methods=["GET"])
-@login_required
-def unread_notifications_count():
-    current = g.current_user["id"]
-    row = query_one(
-        "SELECT COUNT(*) AS c FROM notifications WHERE user_id = ? AND is_read = 0",
-        (current,),
-    )
-    return jsonify({"unread": row["c"] if row else 0})
-
-
-@profile_bp.route("/notifications/mark-read", methods=["POST"])
-@login_required
-def mark_notifications_read():
-    current = g.current_user["id"]
-    execute("UPDATE notifications SET is_read = 1 WHERE user_id = ?", (current,))
-    return jsonify({"ok": True})
