+++ app/templates/notifications.html
@@ -1,0 +1,28 @@
+{% extends "components/base.html" %}
+{% block title %}Notifications - Matcha{% endblock %}
+{% block content %}
+<div class="container py-5">
+    <div class="d-flex justify-content-between align-items-center mb-4">
+        <h1>Notifications</h1>
+        <form method="post" action="{{ url_for('notifications.mark_notifications_read') }}">
+            <button type="submit" class="btn btn-outline-danger">Mark all read</button>
+        </form>
+    </div>
+
+    <ul class="list-group">
+        {% for notif in notifications %}
+        <li class="list-group-item d-flex justify-content-between align-items-center {% if not notif.is_read %}fw-bold{% endif %}">
+            <span>
+                {{ notif.type | replace('_', ' ') | capitalize }}
+                <small class="text-muted d-block">{{ notif.created_at }}</small>
+            </span>
+            {% if not notif.is_read %}
+            <span class="badge bg-danger">New</span>
+            {% endif %}
+        </li>
+        {% else %}
+        <li class="list-group-item text-muted">No notifications yet.</li>
+        {% endfor %}
+    </ul>
+</div>
+{% endblock %}
